Fixed some bugs related to int size

+++ src/sprite.h
@@ -8,8 +8,8 @@ typedef int (*Button)(struct Props *, struct Layer *, short y, short x);
 typedef int (*Hover)(struct Props *, struct Layer *, char start, short y, short x);
 
 typedef struct Sprite {
-	uint8_t colourDepth, iconDepth, buttonDepth, hoverDepth;
-	uint8_t colourMaxDepth, iconMaxDepth, buttonMaxDepth, hoverMaxDepth;
+	uint16_t colourDepth, iconDepth, buttonDepth, hoverDepth;
+	uint16_t colourMaxDepth, iconMaxDepth, buttonMaxDepth, hoverMaxDepth;
 	Colour *colour;
 	char (*icon)[3];
 	Button *button;
